removed obsolete v1 code
