+++ lib/livingstyleguide/filters/full_width.rb
@@ -1,7 +1,3 @@
-LivingStyleGuide::Example.add_filter :full_width do
-  add_wrapper_class '-lsg-has-full-width'
-end
-
 LivingStyleGuide.add_filter :full_width do
   document.classes << '-lsg-has-full-width'
   nil
