--- lib/livingstyleguide/sass_extensions.rb
@@ -1,9 +1,0 @@
-module LivingStyleGuide::SassExtensions
-end
-
-require 'livingstyleguide/sass_extensions/functions'
-require 'sass'
-
-module Sass::Script::Functions
-  include LivingStyleGuide::SassExtensions::Functions
-end
